add cart item

+++ src/main/java/com/example/api_beverage_shop/dto/ToppingDTO.java
@@ -4,5 +4,5 @@ import java.math.BigDecimal;
 
 public class ToppingDTO {
     private String toppingName;
-    private BigDecimal toppingPrice;
+    private Integer toppingPrice;
 }
